+++ apps/user/src/components/common/virtualScroll/virtualScroll.css.ts
@@ -1,0 +1,6 @@
+import { style } from "@vanilla-extract/css";
+
+export const container = style({
+  overflowY: "auto",
+  height: "100%",
+});
